Move endpoints to separate dedicated modules
Currently we have everything in endpoints.rs. This is not a scalable
solution as adding more and more endpoint code to a single file can
make it bigger and difficult to manage. This commit moves endpoints
to their dedicated module. These modules will be move to a new
directory called endpoints which will house all the endpoint modules.