Check activation code length before activating an account
When user enters wrong activation code or a code with unexpected
length, tyto checks in database if code exists. This can create problem
if someone has initiated DoS attach and has sent so many activaiton
retuests with invalid code.

This commit adds a check for specific length in this scenario.